Refactor, lint, types
